+++ connections-app/src/app/core/components/header/header.component.html
