+++ connections-app/src/app/core/components/sidebar/sidebar.component.html
