+++ connections-app/src/app/core/directives/outside-click.directive.ts
